+++ src/components/paginaInicial.js
@@ -1,24 +1,21 @@
 import React from "react";
 import styled from "styled-components";
 import Logo from "../img/Logo.png";
+import { Button, Typography } from '@material-ui/core'
 
 const Head = styled.header`
   display: flex;
-  align-items: center;
-  justify-content: space-between;
-  background-color: #fcebd0;
-  position: fixed;
-  width: 100%;
-  height: 100%;
+  text-align:center;
+  align-items:center;
+  margin-left:600px;
+
 `;
 
 const Titulo = styled.h1`
   width: 100%;
-  margin-bottom: 200px;
   padding: 10px;
-  float: left;
+  margin:0;
   text-align: center;
-  line-height: 50px;
   border-top: 4px solid #5ce;
   background-color: #f3940c;
   color: #ffffff;
@@ -27,26 +24,32 @@ const Titulo = styled.h1`
 `;
 
 const ImagemLogo = styled.img`
-  width: 150px;
-  height: 150px;
-  position: absolute;
-  top: 3%;
-  left: 43%;
+display:flex;
+align-items:center;
+width: 150px;
+height: 150px;
+text-align:center;
+margin:0;
+ 
 `;
 
 class PaginaInicial extends React.Component {
   render() {
     return (
       <div>
-      <Head>
+        <Head>
         <ImagemLogo src={Logo} alt="Logo do Elo4" />
+       
+        </Head>
         <Titulo>Seja muito bem vindo a Elo4 !!!</Titulo>
-        <button onClick={this.props.botaoFornecedor}>Fornecedor</button>
-      <button onClick={this.props.botaoCliente}>Cliente</button>
-      </Head>
-      
-      
-      
+<section>    
+<Button variant='contained' size='medium' color='primary' onClick={this.props.botaoFornecedor}>Fornecedor
+</Button>
+</section>
+<section>
+<Button variant='contained' size='medium' color='primary' onClick={this.props.botaoCliente}>Cliente
+</Button>                   
+</section> 
       </div>
     );
   }
